Set up tooling to dev in docker image
Set up the tools in order to run the commands in a docker container,
as I was not able to install the project on my setup.

After that the tests were failing.

I updated two date-related assertions; looking at the ONIX fixtures, the
new expected values seem to make more sense, though the result might
depend on the system timezone settings or something like that.
It would explain why the tests failed in the container, as before the
result would have depended on the settings of the system running the
tests.